requirements.txt
streamlit==1.36.0
pandas==2.2.2
gspread==6.1.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
openpyxl==3.1.5
